requirements.txt
gunicorn
flask
flask-cors
numpy
openai
azure.cognitiveservices.speech
python-ffmpeg

